Blade server pools (#32)
* Switch test framework to nose

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in ucsmsdk_samples/admin

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in ucsmsdk_samples/firmware

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add .gitignore from ucsmsdk

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in ucsmsdk_samples/network

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in ucsmsdk_samples/reports

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in ucsmsdk_samples/server

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Fix all flake8 errors in tests

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add unit test for port_create functions

This PR adds unit tests for both the
uplink_port_create and the
server_port_create functions

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function to create IPMI policy

This patch adds a function create an IPMI policy
which can be included in a service profile.
Optionally, user authentication can be added
as well.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add unit tests for ipmi functions

This PR adds tests for (in)valid ipmi
policies.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function to set drive state

Before a service profile which includes a local
disk policy can be applied, the disk state
needs to be set in order to accept the policy.
This commit adds a function which allows the
drive state to be set to either "jbod" or
"unconfigured-good".

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add unit tests for setting disk state

This commit adds (in)valid unit tests for
setting the disk state.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add server_pool functions

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add unit tests for server_pool

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add server pool to service profile template

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add PXE option for boot policy

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function to acknowledge chassis

When doing initial fabric setup, until a
chassis is acknowledged, the blades will
not go through discovery.  This commit
adds a function to easily acknowledge
a chassis.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function for adding blades to pool

When the server pool functions were first
created, they only included rackmounts. This
patch also adds blade servers.

Signed-off-by: Britt Houser <bhouser@cisco.com>